services/rag-service/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
qdrant-client
python-jose==3.4.0
python-multipart
openai
pdfminer.six
sqlalchemy
asyncpg
sentence_transformers
httpx
pytest
pytest-asyncio
pytest-cov
psycopg2-binary
black
isort
flake8
mypy
